Update and rename README.md to README .md
a. I never claimed it was mine
b. I never updated my profile like that.
and I'm currently researching data that has used my GitHub account
c. Nowadays, many smart people pretend to be smart when they're just copying things, so I prefer to be stupid and keep quiet
d. Those who always use other people's accounts for malicious repositories are nothing more than losers pretending to be winners, and that only applies in their minds, not to the public and in reality.

The discussion applies to Page 10 and beyond.